JAT - convert image to map; merge maps; basic image and map plotting
See examples/example106*.py for an example of how the LosImage.interp_to_map() function works.

See examples/Create_map_outline.py to see the working outline of how a set of images is converted to a map and saved in the DB.  Specifically, this shows the new combine_maps() function.