JAT - DB schema update; batch download function
The biggest changes in this commit are to the database schema.  These can be seen in modules/DB_classes_v2.py.  These changes propogated through out the code base, especially in DB_funs_v2.py and datatypes_v2.py.

A batch image download function was added in modules/image_download.py.  See example104_download-image-batch.py.  Writing and using this caused me to update error handling in many of the download functions.

Also, finished PsiMap.write_to_file().  This writes to hdf5, but is not compatible with psitools.  psitools checks for the number of groups.  This file has too many, however it would probably be relatively easy to write a wrapper that removes the groups that are unnecessary for psitools.