JAT - add auto datatype coercion to DB scheme
The new addition requires the python package sqlalchemy_utils.  Now any, queries or writes to the database will attempt to auto-coerce data types to an compatible type.  Additionally, if auto-coercion is not supported an exception will be thrown (previously it would write garbage or return no matches to the query).

Not all data types that we are using are supported for auto-coercion.  Specifically, astropy.Time() objects are not.  However, all pandas dataframe types that result from a query are now auto coerced.  So our DB inputs and outputs are now interchangeable.