JT - Add 'instrument' column to 'image_combos' table to optimize queries
The instrument column of image_combos will show an instrument ["AIA", "EUVI-A", "EUVI-B"] if ALL images in the combo are from the same instrument. Otherwise the value will be set to "MULTIPLE" or None.  This dramatically improves image_combo querying time.
- Added function DB_funs.combo_bracket_date() to query the DB directly for the combos before and after the target datetime. This simplifies our combo matching logic and takes negligible time due to the query time improvement.
- Built a series of functions off of this that do one LBCC/IIT coeficient query for each image operated on.  Slightly less effecient, but the returned coefficients are no longer dependent on user entering an appropriate query time-range.
- Everything is backward compatible.  The only function I replaced is DB_funs.query_inst_combo() (to take advantage of new faster querying), but the old version is still available DB_funs.query_inst_combo_old()

Separately, I trimmed the LBCC coefficients to dates where we are more comfortable with the results (as we discussed) using analysis/lbcc_analysis/LBCC_trim-coefs_DB.py